triangulation/requirements.txt
numpy==1.26.2
matplotlib==3.8.2
geopandas==0.14.1
requests==2.31.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
GeoAlchemy2==0.14.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
# tensorflow-macos==2.15.0
tensorflow-object-detection-api==0.1.1
tqdm==4.66.1
joblib==1.3.2
Pillow==10.1.0
pyproj==3.6.1
shapely==2.0.2
psycopg2-binary==2.9.9